changed Map into an interface
